remove Entitlements.plist and don't add to CodeResources
Entitlements.plist is just created temporarily while signing packages
but shouldn't be added to the final output nor should its hash be
included in CodeResources.

We don't add a rule to the _template.xml so that the template can
continue to just match the rules we see from Apple's codesign tool and
instead just add a special case check to avoid scanning the file.