stash: lots of sign-from-scratch fixes/changes
Still not quite working (but seeming very close now) to sign from
scratch and install on iOS 11 and this commit really needs to be split
up but I've nearly run out of steam looking at this :/

If signing from scratch isign will now create a second CodeDirectory
with sha256 hashes.

The resource rules are now stored in-order within code_resources.py
instead of the code_resources_template.xml to avoid depending on any
undefined sorting behaviour of the plist parser (which was noted to be
different between python2/3)

Note this was based on a commit before 2to3 was run to adding python3
compatibility since after it was looking like makesig code was failing
due to python3 compatibility issues. The plan was to bring back the
python3 support again only after getting it working with python2.

To support python2 this went back to using biblist to parse binary
plists